requirements.txt
fastmcp==2.10.6
mcp==1.12.0
openpyxl==3.1.5
reportlab==4.4.2
twilio==9.6.5
pydantic==2.11.7
python-dotenv==1.1.1
uvicorn==0.35.0